+++ Lilu/Headers/kern_patcher.hpp
@@ -222,6 +222,73 @@ public:
 
 		return (T)nullptr;
 	}
+    
+    /**
+     *  Solve request to resolve multiple symbols in one shot and simplify error handling
+     *
+     *  @seealso solveMultiple().
+     */
+    struct SolveRequest {
+        /**
+         *  The symbol to solve
+         */
+        const char *symbol {nullptr};
+        
+        /**
+         *  The symbol address on success, otherwise NULL.
+         */
+        mach_vm_address_t *address {nullptr};
+        
+        /**
+         *  Construct a solve request conveniently
+         */
+        template <typename T>
+        SolveRequest(const char *s, T &addr) :
+			symbol(s), address(reinterpret_cast<mach_vm_address_t*>(&addr)) { }
+    };
+	
+	/**
+	 *  Solve multiple functions with basic error handling
+	 *
+	 *  @param id        loaded kinfo id
+	 *  @param requests  an array of requests to solve
+	 *  @param num       requests array size
+	 *  @param start     start address range
+	 *  @param size      address range size
+	 *  @param crash     kernel panic on invalid non-zero address
+	 *  @param force     continue on first error
+	 *
+	 *  @return false if at least one symbol cannot be solved.
+	 */
+	inline bool solveMultiple(size_t id, SolveRequest *requests, size_t num, mach_vm_address_t start, size_t size, bool crash=false, bool force=false) {
+		for (size_t index = 0; index < num; index++) {
+			auto result = solveSymbol(id, requests[index].symbol, start, size, crash);
+			if (result) {
+				*requests[index].address = result;
+			} else {
+				clearError();
+				if (!force) return false;
+			}
+		}
+		return true;
+	}
+	
+	/**
+	 *  Solve multiple functions with basic error handling
+	 *
+	 *  @param id        loaded kinfo id
+	 *  @param requests  an array of requests to solve
+	 *  @param start     start address range
+	 *  @param size      address range size
+	 *  @param crash     kernel panic on invalid non-zero address
+	 *  @param force     continue on first error
+	 *
+	 *  @return false if at least one symbol cannot be solved.
+	 */
+	template <size_t N>
+	inline bool solveMultiple(size_t id, SolveRequest (&requests)[N], mach_vm_address_t start, size_t size, bool crash=false, bool force=false) {
+		return solveMultiple(id, requests, N, start, size, crash, force);
+	}
 
 	/**
 	 *  Hook kext loading and unloading to access kexts at early stage
@@ -407,6 +474,16 @@ public:
 		template <typename T>
 		RouteRequest(const char *s, T t, mach_vm_address_t &o) :
 			symbol(s), to(reinterpret_cast<mach_vm_address_t>(t)), org(&o) { }
+		
+		/**
+		 *  Construct RouteRequest for wrapping a function
+		 *  @param s  symbol to lookup
+		 *  @param t  destination address
+		 *  @param o  trampoline storage to the original symbol
+		 */
+		template <typename T, typename O>
+		RouteRequest(const char *s, T t, O &o) :
+			RouteRequest(s, t, reinterpret_cast<mach_vm_address_t&>(o)) { }
 
 		/**
 		 *  Construct RouteRequest for routing a function
@@ -514,6 +591,29 @@ public:
 		return routeMultipleShort(id, requests, N, start, size, kernelRoute, force);
 	}
 
+	/**
+	 *  Simple find and replace in kernel memory.
+	 */
+	static inline bool findAndReplace(void *data, size_t dataSize, const void *find, size_t findSize, const void *replace, size_t replaceSize) {
+		void *res;
+		if (UNLIKELY((res = lilu_os_memmem(data, dataSize, find, findSize)) != nullptr)) {
+			if (UNLIKELY(MachInfo::setKernelWriting(true, KernelPatcher::kernelWriteLock) != KERN_SUCCESS)) {
+				SYSLOG("patcher", "failed to obtain write permissions for f/r");
+				return false;
+			}
+
+			lilu_os_memcpy(res, replace, replaceSize);
+
+			if (UNLIKELY(MachInfo::setKernelWriting(false, KernelPatcher::kernelWriteLock) != KERN_SUCCESS)) {
+				SYSLOG("patcher", "failed to restore write permissions for f/r");
+			}
+
+			return true;
+		}
+
+		return false;
+	}
+
 private:
 	/**
 	 *  Jump type for routing
@@ -548,11 +648,12 @@ private:
 	/**
 	 *  Read previous jump destination from function
 	 *
-	 *  @param from         formerly routed function
+	 *  @param from          formerly routed function
+	 *  @param jumpType previous jump type
 	 *
 	 *  @return wrapper pointer on success or 0
 	 */
-	mach_vm_address_t readChain(mach_vm_address_t from);
+	mach_vm_address_t readChain(mach_vm_address_t from, JumpType &jumpType);
 
 	/**
 	 *  Created routed trampoline page
@@ -600,27 +701,35 @@ private:
 
 #ifdef LILU_KEXTPATCH_SUPPORT
 	/**
-	 *  Called at kext loading and unloading if kext listening is enabled
-	 */
-	static void onKextSummariesUpdated();
-
-	/**
-	 *  A pointer to loaded kext information
-	 */
-	OSKextLoadedKextSummaryHeaderAny **loadedKextSummaries {nullptr};
-
-	/**
-	 *  A pointer to kext summaries update
-	 */
-	void (*orgUpdateLoadedKextSummaries)(void) {nullptr};
-
-	/**
 	 *  Process already loaded kexts once at the start
 	 *
-	 *  @param summaries loaded kext summaries
-	 *  @param num       number of loaded kext summaries
 	 */
-	void processAlreadyLoadedKexts(OSKextLoadedKextSummaryHeaderAny *summaries, size_t num);
+	void processAlreadyLoadedKexts();
+
+	/**
+	 *  Pointer to loaded kmods for kexts
+	 */
+	kmod_info_t **kextKmods {nullptr};
+
+	/**
+	 *  Called at kext unloading if kext listening is enabled
+	 */
+	static OSReturn onOSKextUnload(void *thisKext);
+
+	/**
+	 *  A pointer to OSKext::unload()
+	 */
+	mach_vm_address_t orgOSKextUnload {};
+
+	/**
+	 *  Called at kext loading and unloading if kext listening is enabled
+	 */
+	static void onOSKextSaveLoadedKextPanicList();
+
+	/**
+	 *  A pointer to OSKext::saveLoadedKextPanicList()
+	 */
+	mach_vm_address_t orgOSKextSaveLoadedKextPanicList {};
 
 #endif /* LILU_KEXTPATCH_SUPPORT */
 
@@ -649,6 +758,11 @@ private:
 	 *  Awaiting already loaded kext list
 	 */
 	bool waitingForAlreadyLoadedKexts {false};
+
+	/**
+	 *  Flag to prevent kext processing during an unload
+	 */
+	bool isKextUnloading {false};
 
 #endif /* LILU_KEXTPATCH_SUPPORT */