+++ Lilu/Sources/kern_start.cpp
@@ -16,23 +16,13 @@
 #include <Headers/kern_cpu.hpp>
 #include <Headers/kern_file.hpp>
 #include <Headers/kern_time.hpp>
+#include <Headers/kern_version.hpp>
 
 #include <IOKit/IOLib.h>
 #include <IOKit/IORegistryEntry.h>
 #include <mach/mach_types.h>
 
 OSDefineMetaClassAndStructors(PRODUCT_NAME, IOService)
-
-static const char kextVersion[] {
-#ifdef DEBUG
-	'D', 'B', 'G', '-',
-#else
-	'R', 'E', 'L', '-',
-#endif
-	xStringify(MODULE_VERSION)[0], xStringify(MODULE_VERSION)[2], xStringify(MODULE_VERSION)[4], '-',
-	getBuildYear<0>(), getBuildYear<1>(), getBuildYear<2>(), getBuildYear<3>(), '-',
-	getBuildMonth<0>(), getBuildMonth<1>(), '-', getBuildDay<0>(), getBuildDay<1>(), '\0'
-};
 
 IOService *PRODUCT_NAME::probe(IOService *provider, SInt32 *score) {
 	setProperty("VersionInfo", kextVersion);
@@ -229,7 +219,8 @@ bool Configuration::getBootArguments() {
 
 	betaForAll = checkKernelArgument(bootargBetaAll);
 	debugForAll = checkKernelArgument(bootargDebugAll);
-	isUserDisabled = checkKernelArgument(bootargUserOff) || getKernelVersion() >= KernelVersion::BigSur;
+	isUserDisabled = checkKernelArgument(bootargUserOff) ||
+		getKernelVersion() <= KernelVersion::SnowLeopard || getKernelVersion() >= KernelVersion::BigSur;
 
 	PE_parse_boot_argn(bootargDelay, &ADDPR(debugPrintDelay), sizeof(ADDPR(debugPrintDelay)));